+++ src/main/java/de/automatic/ui/colorconfiguraiton/entities/Sample.java
@@ -2,17 +2,17 @@ package de.automatic.ui.colorconfiguraiton.entities;
 
 public interface Sample {
 
-	public int get(Channels c);
+	public double get(Channels c);
 	
-	public int getC1();
+	public double getC1();
 
 	public void setC1(int r);
 
-	public int getC2();
+	public double getC2();
 
 	public void setC2(int g);
 
-	public int getC3();
+	public double getC3();
 
 	public void setC3(int b);
